% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonMatching.r
\name{\%!in\%}
\alias{\%!in\%}
\title{Value Non-Matching (inverse of \%in\% )}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched}

\item{y}{vector or \code{NULL}: the values to be matched against}
}
\description{
returns a vector of the positions of (first) non-matches of its first argument in its second.
}
\details{
inverse of \%in\%
}
