% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idTEsProt.r
\docType{data}
\name{idTEsProt}
\alias{idTEsProt}
\title{Genes related to transposable elements proteins used in the cleanRep.cds function}
\format{
vector
}
\source{
UniProt
}
\usage{
data(idTEsProt)
}
\description{
Genes related to transposable elements proteins used in the cleanRep.cds function
}
\keyword{datasets}
