% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.RTEs.r
\name{import.RTEs}
\alias{import.RTEs}
\title{Loads retrotransposons counts and return a DGEList object for use with edgeR.}
\usage{
import.RTEs(path.sal, ref.sal, conditions, import_to = c("edgeR, DESeq2"))
}
\arguments{
\item{path.sal}{output salmon directory}

\item{ref.sal}{salmon reference build with mk.reference function}

\item{conditions}{A string vector indicating the experimental conditions of each library. Example: c("S1", "S1", "S1", "S2", "S2", "S2", "S3", "S3", "S3")}

\item{import_to}{A string object indicating whether to import for edgeR or DESeq2. If none is selected, transcript-level estimates are returned.}
}
\description{
Imports count estimates and calculates an offset that corrects for changes to the average transcript length across samples (details in https://bioconductor.org/packages/release/bioc/vignettes/tximport/inst/doc/tximport.html#edgeR). The function returns a DGEList for use in downstream analysis with edgeR
}
