% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk.reference_mo.r
\name{mk.reference_mo}
\alias{mk.reference_mo}
\title{Makes a reference file for Salmon in model organisms}
\usage{
mk.reference_mo(
  bedtools = "bedtools",
  GFFgen,
  genome,
  trme,
  RMgen,
  RMtr,
  overlapping = T,
  by = c("namRep", "classRep", "class", "supFam", "Fam"),
  threads = 1,
  outdir,
  over.res = c("HS", "LE", "LD"),
  ...
)
}
\arguments{
\item{bedtools}{bedtools binary path}

\item{GFFgen}{gtf file to reference genome}

\item{genome}{reference genome in fasta format}

\item{trme}{transcriptome in fasta format}

\item{RMgen}{RepeatMasker output file from genome}

\item{RMtr}{RepeatMasker output file from transcriptome}

\item{outdir}{output directory}
}
\description{
This function creates decoys and a transcriptome that will be used by Salmon when the reference genome is available.
}
