% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ovlp.res.r
\name{ovlp.res}
\alias{ovlp.res}
\title{Overlappings resolution}
\usage{
ovlp.res(
  RepMask,
  anot,
  gff3,
  stranded = T,
  outdir,
  rm.cotrans = F,
  align,
  threads = 1,
  ignore.aln.pos = T,
  over.res = c("HS", "LS", "LD"),
  ...
)
}
\arguments{
\item{RepMask}{RepeatMasker output file. If rm.cotrans = F, them you must enter a RepeatMasker output file without co-transcripted repeats.}

\item{anot}{annotation file in outfmt6 format. It is necessary when the option rm.cotrans = T}

\item{gff3}{gff3 file. It is necessary when the option rm.cotrans = T}

\item{stranded}{logical vector indicating if the library is strand specific}

\item{outdir}{Output directory}

\item{align}{.align file}

\item{threads}{Number of cores to use in the processing. By default threads = 1}

\item{ignore.aln.pos}{The RepeatMasker alignments file may have discrepancies in the repeats positions with respect to the output file. If you selected over.res = "LD", then you can choose whether to take into account the positions of the alignment file or to take the average per repeats class (default).}

\item{over.res}{Indicates the method by which the repetition overlap will be resolved ("HS" by default).
HS: higher score, bases are assigned to the element with the highest score
LS: longer element, bases are assigned to the longest element
LD: lower divergence, bases are assigned to the element with the least divergence.
in all cases both elements have the same characteristics, the bases are assigned to the first element.}

\item{rm.cotrnas}{logical vector indicating whether co-transcribed repeats should be removed}

\item{cleanTEsProt}{logical vector indicating whether the search for TEs-related proteins should be carried out (e.g.
transposases, integrases, env, reverse transcriptase, etc.). We recommend that users use a curated annotations file,
in which these genes have been excluded; therefore the default option is F. When T is selected, a search is performed
against a database obtained from UniProt, so we recommend that the annotations file have this format for the subject
sequence id (e.g. "CO1A2_MOUSE"/"sp|Q01149|CO1A2_MOUSE"/"tr|H9GLU4|H9GLU4_ANOCA")}

\item{featureSum}{Returns statistics related to the characteristics of the transcripts. Requires a gff3 file. If TRUE, returns a list of the}
}
\description{
This function resolves overlapping repeats assigned to the same transcript and returns a data frame of repeats with no overlaps. The user can define the criteria to solve the overlaps, either by higher score (HS), longer length (LE) or lower Kimura's distances (LD).
}
