% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.alignfile.r
\name{read.alignfile}
\alias{read.alignfile}
\title{Read .align file from RepeatMasker}
\usage{
read.alignfile(file, average = F, by = c("seqID", "classRep"))
}
\arguments{
\item{file}{.align file}

\item{average}{logical vector indicating if the mean should be taken}

\item{by}{if \code{average = TRUE}, then it must be indicated if the average is taken by sequences, \code{by = "seqID"}, or by repetition class, \code{by = "classRep"}.}
}
\description{
read .align file from RepeatMasker
}
\details{
reads the RepeatMasker .align file and returns a data.frame with the sequence ID, the repetitions class, and the sequence divergence respect to the reference sequence in Kimura 2-parameter distances. If \code{average = TRUE} return a vector with means by sequences (\code{by = "seqID"}) or by repetition class (\code{by = "classRep"}) as indicated.
}
