% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.gff3.r
\name{read.gff3}
\alias{read.gff3}
\title{Read gff3 file}
\usage{
read.gff3(file)
}
\arguments{
\item{file}{gff3 file}
}
\description{
reads a gff3 file from TransDecoder output
}
\details{
This function read an gff3 file from TransDecoder output and rename columns as:
\describe{
\item{seqid}{sequence ID}
\item{source}{algorithm or operating procedure that generated the feature}
\item{feature}{The type of the feature}
\item{strt}{The start of the feature}
\item{end}{The end of the feature}
\item{score}{The score of the feature}
\item{st}{strand (+ or -)}
\item{phase}{For features of type "CDS", the phase indicates where the feature begins with reference to the reading frame.}
\item{att}{attributes}
}
}
