% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.outfmt6.r
\name{read.outfmt6}
\alias{read.outfmt6}
\title{Read outfmt6 file}
\usage{
read.outfmt6(file)
}
\arguments{
\item{file}{outfmt6 file}
}
\description{
reads a annotations file in output format 6
}
\details{
This function read a annotations file in output format 6 and rename columns as:
\describe{
\item{seqid}{query (e.g., unknown gene) sequence id}
\item{sseqid}{subject (e.g., reference genome) sequence id}
\item{pident}{percentage of identical matches}
\item{length}{alignment length (sequence overlap)}
\item{mismatch}{number of mismatches}
\item{gapopen}{ number of gap openings}
\item{qstart}{start of alignment in query}
\item{qend}{end of alignment in query}
\item{sstart}{start of alignment in subject}
\item{send}{end of alignment in subject}
\item{evalue}{expect value}
\item{bitscore}{bit score}
}
}
