% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readalin_sh.r
\docType{data}
\name{readalin_sh}
\alias{readalin_sh}
\title{shell script to read .align and .align.gz files}
\format{
An object of class \code{character} of length 1.
}
\source{
bash
}
\usage{
data(readalin_sh)
}
\description{
shell script to read .align and .align.gz files
}
\keyword{datasets}
