% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.salmon.r
\name{run.salmon}
\alias{run.salmon}
\title{Run Salmon}
\usage{
run.salmon(
  index,
  decoys,
  salmon_path = "salmon",
  kmer,
  lib_dir,
  pe_se = c("pe", "se"),
  threads,
  trme
)
}
\arguments{
\item{index}{paterrn of output salmon directory. This tag will be used to create the index and a folder containing the counts for each library}

\item{decoys}{The location path of the \code{decoy.txt} file generated by the \code{mk.reference} function}

\item{salmon_path}{A path indicating the location of the program}

\item{kmer}{The kmer size used during Salmon indexing}

\item{lib_dir}{A path indicating the location of libraries. Libraries should be tagged with ".fastq.gz" (single-end) or "_R#.fastq.gz" (paired-end) at the end, for example: anyName.fastq.gz (single-end) or anyName_R1.fastq.gz and anyName_R2.fastq.gz (paired-end)}

\item{pe_se}{Indicate if you using paired-end (\code{"pe"}) or single-end (\code{"se"}) libraries}

\item{threads}{number of threads to be used}

\item{trme}{A path indicating the location of transcriptome}

\item{output_dir}{output salmon directory}
}
\description{
This function helps the user to run the Salmon program. The user must assign the location of the program, an output directory, the transcriptome, the location of the libraries, the kmer size, the processing threads, and a label for the index.
}
