% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.vec.r
\name{split.vec}
\alias{split.vec}
\title{Split vector}
\usage{
split.vec(vec, pattern, fragPos = 1)
}
\arguments{
\item{vec}{vector to split}

\item{pattern}{character that indicates the pattern by which the vector will be divided}

\item{fragPos}{the position of the fragment to return}
}
\description{
Splits a vector according to a pattern into fragments and returns the fragment at the fragPos position
}
