% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l.WR.r
\name{Wickerlike.rule}
\alias{Wickerlike.rule}
\title{Applies a rule-based sequence selection criterion similar to Wicker's rule on processed transcripts/sequences.}
\usage{
Wickerlike.rule(
  RepMask,
  by = "classRep",
  rule = c(80, 80, 80),
  best.by = NULL,
  custom.lengths = NULL
)
}
\arguments{
\item{RepMask}{RepeatMasker output file.}

\item{by}{The column by which the repeats will be classified: 'classRep' (default) or 'namRep'.}

\item{rule}{A numerical vector respectively indicating the minimum percentage of identity, the percentage of the length of class/family repeat with respect to the length of the transcript, and the length (in base pairs) of the repeat to be analyzed.
#The position of the numbers indicates respectively:
Example: c(80, 60, 100) indicates that those repeats with 80\% identity or more in at least 60\% of the transcript, and are at least 100 bp in length will be annotated as target TEs.  Default is c(80,80,80)}

\item{best.by}{Defines if only the best match of each transcript/sequence id should be returned (by default best.by = NULL which shows all matches for the sequence). The user can choose whether to be based on the longest repeat length ('total_repeat_length') or the highest percent identity ('per_identity').
A logical vector indicating its only the longest repeats for each transcript is reported. By default best = TRUE}

\item{custom.lengths}{A a numerical vector indicating the lengths for each transcript. By default lengths are calculated from the RepeatMasker file.}
}
\description{
Applies a rule-based sequence selection criterion similar to Wicker's rule on processed transcripts / sequences. For each sequence identification, take the percent identity (100 minus the percent divergence), the ratio of the TE identified in the sequence respect to the total sequence length, and the minimum transcript length. The function returns a table with the sequence IDs that meet the criteria, the repeats name and its positions within the sequence, and other auxiliary variables (percentage of identity, number of total bases for each repetition and length of the transcript). The user can define whether to summarize all the matches found in each sequence or only the best match.
}
