% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk.reference.r
\name{mk.reference}
\alias{mk.reference}
\title{Makes a reference file for Salmon}
\usage{
mk.reference(
  RepMask,
  overlapping = F,
  by = c("namRep", "classRep", "class", "supFam", "Fam"),
  rule = c(80, 80, 80),
  best = T,
  trme,
  threads = 1,
  outdir,
  ...
)
}
\arguments{
\item{RepMask}{RepeatMasker output file. If rm.cotrans = F it is assumed that the file does not contain cotranscribed repeats. If overlapping = F it is assumed that the file does not contain overlapping.}

\item{overlapping}{Indicates whether the RepMask file contains overlapping repetitions (TRUE) or not (FALSE). When the RepMask file contains overlapping repetitions, the ovlp.res() function will be used to solve them and the resolution criteria must be indicated (higher score (HS), longer length (LE) or lower Kimura distances (LD))}

\item{by}{The column by which the repeats will be classified}

\item{rule}{A numerical vector respectively indicating the minimum percentage of identity, the percentage of the minimum length of a repeat with respect to the length of the transcript, and the length in base pairs of the repeats to be analyzed. Example: c(80, 80, 80)}

\item{best}{If best = T, the repetition with the highest percentage of coverage in the transcript will be selected (by default).}

\item{trme}{transcriptome in fasta format}

\item{threads}{Number of cores to use in the processing. By default threads = 1}

\item{outdir}{Output directory}

\item{rm.cotrnas}{logical vector indicating whether co-transcribed repeats should be removed}

\item{align}{.align file}

\item{over.res}{Indicates the method by which the repetition overlap will be resolved.
HS: higher score, bases are assigned to the element with the highest score
LS: longer element, bases are assigned to the longest element
LD: lower divergence, bases are assigned to the element with the least divergence.
in all cases both elements have the same characteristics, the bases are assigned to the first element.}

\item{anot}{annotation file in outfmt6 format. It is necessary when the option rm.cotrans = T}

\item{gff3}{gff3 file. It is necessary when the option rm.cotrans = T}

\item{stranded}{logical vector indicating if the library is strand specific}

\item{cleanTEsProt}{logical vector indicating whether the search for TEs-related proteins should be carried out (e.g.
transposases, integrases, env, reverse transcriptase, etc.). We recommend that users use a curated annotations file,
in which these genes have been excluded; therefore the default option is F. When T is selected, a search is performed
against a database obtained from UniProt, so we recommend that the annotations file have this format for the subject
sequence id (e.g. "CO1A2_MOUSE"/"sp|Q01149|CO1A2_MOUSE"/"tr|H9GLU4|H9GLU4_ANOCA")}

\item{featureSum}{Returns statistics related to the characteristics of the transcripts. Requires a gff3 file. If TRUE, returns a list of the}

\item{ignore.aln.pos}{The RepeatMasker alignments file may have discrepancies in the repeats positions with respect to the output file. If you selected over.res = "LD", then you can choose whether to take into account the positions of the alignment file or to take the average per repeats class (default).}
}
\description{
This function creates a reference file that will be used by Salmon to create an index from a RepMask file.
The user can enter a RepMask file without deleting co-transcribed or overlapping repeats with the RepMask argument, or enter a RepMask file without co-transcribed but overlapping repeats with the RepMask.clean argument, or a file free of co-transcribed or overlapping repeats with the RepMask.ovlp.clean argument. When the file contains co-transcribed repeats, it must indicate rm.cotrans = T and when the file contains overlaps it must indicate overlapping = T.
}
